review: keep one copy of the env, descriptor and digest helpers
lastEnvValue and parseDescriptor were byte-identical in the cocoon and
process engines and imageDigest differed only by cocoon's url check.
The env helper joins utils; the image ones join SplitRef in engine/types,
where they need no new package dependency, and process picks up the url
form with them.
